EOS, HeatCool: only compute deltaT once
When calculating heating/cooling terms, the parameter "deltaT" was being
calculated each time one of those terms was computed. However the value is a
runtime constant, so we should compute it only once and store it in
atomic_rates_module. This also saves us a lot of unnecessary divides, which are
expensive.