EOS: change repeated divides to multiplies in the denominator in ion_n()
Compilers may be smart enough to do this transformation automatically, but we
shouldn't trust them. Writing the terms this way guarantees that we will do 3
fewer divides in ion_n(), which is already the most expensive routine in all of
Nyx, thanks to all the divides and the dlog10() term.

Note that changing the order of operations like this *will* change the results
very slightly (especially because the EOS is called so frequently), so we may
need to update the regression test results to accommodate this.