HeatCool: make "e_in" dummy argument in f_rhs_vec() more in line with its behavior in f_rhs()
f_rhs() sets e_in to TINY() if it enters the function with a negative value. In
the original version of f_rhs_vec() the value of e_in() was copied to a local
variable "e" which was then set to TINY if negative. But it is less confusing
for the two functions to have the same behavior.