EOS: add BIND(C) attribute to interp_to_this_z() function, and add C++ function prototype
This function modifies public variables stored in modules, so we want to have
control over which threads execute this function; if multiple threads execute
it simultaneously, then there is a race condition. The OpenMP semantics should
be in the C++ code, not Fortran, so we will move calls to this function out of
existing Fortran functions and into C++ code instead.