Nyx::just_the_hydro(): remove manual initialization of reduction variables
In commit f3bdfe4 ("Nyx::just_the_hydro(): add OpenMP reduction construct to
e_added,ke_added") I moved initialization of the reduction variables "e_added"
and "ke_added" to inside the OpenMP parallel region. However, while this code
is not wrong, it is unnecessary because the OpenMP spec requires that the
default initializer for the "+" reduction is zero, which is what this code
already does by hand. So we can remove it.