EOS, HeatCool: implement SIMD ODE integration (v2) (#20)
* EOS: add SIMD function nyx_eos_given_RT_vec()

This is a SIMD version of nyx_eos_given_RT()

* EOS: add SIMD function ion_n_vec()

This is a SIMD version of the function ion_n()

* EOS: add SIMD function iterate_ne_vec()

This is a SIMD version of the function iterate_ne().

* EOS: add SIMD function nyx_eos_T_given_Re_vec()

This is a SIMD version of the function nyx_eos_T_given_Re().

* Src_3d: added SIMD function fort_compute_temp_vec()

This is a SIMD version of the function fort_compute_temp()

* Initialization: add Nyx::set_simd_width() and Fortran set_simd() functions

These set the SIMD width.

* EOS: add subroutine iterate_ne_vec()

This is a SIMD version of the function iterate_ne().

* EOS: add function nyx_eos_T_given_Re_vec()

This is a SIMD version of the function nyx_eos_T_given_Re().

* EOS: add subroutine hc_rates_vec()

This is a SIMD version of the subroutine hc_rates().

* HeatCool, LyA: add functions f_rhs_vec() and RhsFn_vec()

These are SIMD versions of the functions f_rhs() and RhsFn().

* HeatCool: add array versions of misc VODE variables

These will only be used with the SIMD functions.

* Src_3d: add subroutine fort_compute_temp_vec()

This is a SIMD version of the subroutine fort_compute_temp().

* LyA: add subroutine fcvode_wrapper_vec()

This is a SIMD version of the subroutine fcvode_wrapper().

* LyA: add subroutine integrate_state_fcvode_vec()

This is a SIMD version of the function integrate_state_fcvode().

* Nyx::read_params(): add a few sanity checks for the SIMD CVODE solver

* EOS: replace missing TCOOLMIN_R and TCOOLMAX_R that got dropped during merge

These are the real values (not log10 values) of TCOOLMIN and TCOOLMAX.

* Nyx: print "SIMD CVODE" to STDOUT when using SIMD CVODE integrator

* HeatCool, EOS: revert back to original SIMD versions of compute_temp

* LyA: fix two memory leaks in integrate_state_fcvode_vec()

* LyA: minor formatting fixes; no functional changes

* LyA: get the # of eqns for SIMD CVODE from the "simd_width" parameter

* CVODE: make misc SIMD functions public instead of protected

We will need to call them from main() instead of within a Nyx object (due to
problems with Nyx objects going out of scope during a restart), so we need them
to be public.

* CVODE: allocate SIMD vectors in main() instead of in Nyx member functions

During a checkpoint restart, there were problems with Nyx objects going out of
scope and so these arrays were in the wrong allocation state.

* CVODE: fix a few typos being printed to STDOUT

* EOS: don't assign vec_count at declaration

It gets initialized in the function body, so no reason to initialize it at
declaration.

* EOS: remove commented-out print statement

* HeatCool: set length of dummy arrays explicitly in hc_rates_vec()

Since we know what it is, there is no reason to make the compiler guess.

* HeatCool: make "e_in" dummy argument in f_rhs_vec() more in line with its behavior in f_rhs()

f_rhs() sets e_in to TINY() if it enters the function with a negative value. In
the original version of f_rhs_vec() the value of e_in() was copied to a local
variable "e" which was then set to TINY if negative. But it is less confusing
for the two functions to have the same behavior.

* HeatCool: add SAVE attribute to the SIMD versions of the misc vars in vode_aux

This way they have the same behavior as their scalar counterparts.

* Initialization: allocate vode_aux arrays in an OpenMP parallel region

These are threadprivate variables, so we need to allocate them in a parallel
region or else they will be allocated only on the master thread.

* LyA, CVODE: switch from dense solver to diagonal solver for SIMD COVDE

The Jacobian matrix of a "system" of independent equations is diagonal, so
let's use that to avoid computing all the off-diagonal elements which are all
zero anyway.

* comoving_a_post_restart(): add missed check for SIMD CVODE mode (heat_cool_type=7)

* EOS: make interpolation err msg in iterate_ne()/iterate_ne_vec() more descriptive

* EOS, HeatCool, VODE: make all tolerance parameters run-time inputs

The tolerances used in the heating/cooling integration affect results
differently depending on whether one uses the scalar or SIMD implementation.
Since it will take some work to find tolerances that yield the same result in
both approaches, it will be convenient to be able to specify these at run time. These tolerances are
specified by three parameters:

  - nyx.eos_nr_eps: tolerance for computing Newton-Raphson iteration in
    iterate_ne()
  - nyx.vode_rtol: relative tolerance in VODE integration
  - nyx.vode_atol_scaled: absolute tolerance in VODE integration, scaled by the
    initial value of the independent variable in the ODE

* HeatCool: add code path for SIMD CVODE in ext_src_hc()

Missed this code path earlier when adding the SIMD CVODE runtime option.

* Nyx::read_params(): print the percent sign "%" the correct way

* Nyx::read_params(): add warning message about experimental nature of SIMD CVODE

* HeatCool, CVODE: fix case of uninitialized bool "hot" in SIMD CVODE

* LyA, CVODE: fix SIMD function name in error messages

* EOS: add stubs for SIMD functions if called without USE_CVODE=TRUE and USE_HEATCOOL=TRUE

* CVODE: don't compile CVODE-relevant code unless USE_CVODE=TRUE