HeatCool: document our way out of the simd_width+blocking problem
When we use the SIMD ODE integration method, we loop over the x-direction of
each box/tile in strides of simd_width. However, if the number of cells in the
x-direction is not a multiple of simd_width, we fall off the end of the box and
compute garbage.

One way to fix this is to make simd_width adaptive so that, if the modulus of
the x-length and simd_width is not zero, it shortens itself on the last
iteration so that it stays within the box. However, doing so would require
tearing down the existing CVODE solver scratch memory and reinitializing it
with the shorter size to do the last iteration of the x-direction, then tear
down the solver memory again and reinitialize it again with the original
simd_width. This extra work would likely destroy any speedup that we otherwise
gain from using the SIMD solver method over the scalar cell-by-cell method. So
instead, we simply quit with an error if the modulus is not zero, and require
the user to build boxes that are appropriately sized.